Actualizacion a nuevo server y cambios radicales en como funciona todo (#8)
* Agregados authenticator, provider y syncAdapter

* Actualizacion a nuevo server

* Cambio version playstore

+++ app/src/main/java/com/ort/smartacc/fragment/SearchResultsFragment.java
@@ -73,6 +73,6 @@ public class SearchResultsFragment extends Fragment {
     }
 
     public interface OnRecetaSelectedCallback{
-        void onRecetaSelected(int id);
+        void onRecetaSelected(long id);
     }
 }
